requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
playwright>=1.40.0
nest-asyncio>=1.5.8
pydantic>=2.0.0
python-dotenv>=1.0.0
